+++ Cesium3DTilesSelection/test/TestTilesetContentManager.cpp
@@ -1309,7 +1309,6 @@ TEST_CASE("Test the tileset content manager's post processing for gltf") {
                 overlayTile.getRectangle(),
                 {},
                 {},
-                {},
                 true});
       }
     };